Split `telegram_webhooks_controller_spec` into module-specific files
